Wire basis templates into single obs mapmaker
